requirements.txt
opencv-python==4.7.0.72
face-recognition==1.3.0
dlib==19.24.0
streamlit==1.22.0
mysql-connector-python==8.0.33
numpy==1.23.5
pillow==9.5.0
requests==2.29.0